+++ backend/src/test/java/wooteco/prolog/Documentation.java
@@ -5,20 +5,42 @@ import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.MediaType;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.RestDocumentationExtension;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ActiveProfiles;
+import wooteco.prolog.login.GithubResponses;
+import wooteco.prolog.login.application.dto.TokenRequest;
+import wooteco.prolog.login.application.dto.TokenResponse;
 
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
 
+@ActiveProfiles({"docu", "test"})
 @ExtendWith(RestDocumentationExtension.class)
-public class Documentation extends AcceptanceTest{
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class Documentation {
+    @LocalServerPort
+    private int port;
     protected RequestSpecification spec;
+
+    public TokenResponse 로그인_사용자;
 
     @BeforeEach
     public void setUp(RestDocumentationContextProvider restDocumentation) {
-        setUp();
+        RestAssured.port = port;
+
+        로그인_사용자 = RestAssured.given().log().all()
+                .body(new TokenRequest(GithubResponses.소롱.getCode()))
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/login/token")
+                .then().log().all()
+                .extract().body().as(TokenResponse.class);
 
         this.spec = new RequestSpecBuilder()
                 .addFilter(documentationConfiguration(restDocumentation))
