+++ backend/src/test/java/wooteco/prolog/post/documentation/PostDocumentation.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import wooteco.prolog.Documentation;
+import wooteco.prolog.login.GithubResponses;
 import wooteco.prolog.mission.application.dto.MissionRequest;
 import wooteco.prolog.mission.application.dto.MissionResponse;
 import wooteco.prolog.post.application.dto.PostRequest;
@@ -116,7 +117,7 @@ public class PostDocumentation extends Documentation {
         given("post/mine")
                 .header("Authorization", "Bearer " + 로그인_사용자.getAccessToken())
                 .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/members/{username}/posts", MEMBER.getGithubUserName());
+                .when().get("/members/{username}/posts", GithubResponses.소롱.getLogin());
     }
 
     private void 포스트를_삭제한다(String location) {
